+++ scico/optimize/_admm.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2020-2025 by SCICO Developers
+# Copyright (C) 2020-2026 by SCICO Developers
 # All rights reserved. BSD 3-clause License.
 # This file is part of the SCICO package. Details of the copyright and
 # user license can be found in the 'LICENSE' file distributed with the
@@ -20,12 +20,14 @@ from scico.numpy import Array, BlockArray
 from scico.numpy.linalg import norm
 
 from ._admmaux import (
-    FBlockCircularConvolveSolver,
-    G0BlockCircularConvolveSolver,
     GenericSubproblemSolver,
     LinearSubproblemSolver,
     MatrixSubproblemSolver,
     SubproblemSolver,
+)
+from ._admmauxconv import (
+    FBlockCircularConvolveSolver,
+    G0BlockCircularConvolveSolver,
 )
 from ._common import Optimizer
