+++ scico/optimize/admm.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2022-2023 by SCICO Developers
+# Copyright (C) 2022-2026 by SCICO Developers
 # All rights reserved. BSD 3-clause License.
 # This file is part of the SCICO package. Details of the copyright and
 # user license can be found in the 'LICENSE' file distributed with the
@@ -15,6 +15,8 @@ from ._admmaux import (
     GenericSubproblemSolver,
     LinearSubproblemSolver,
     MatrixSubproblemSolver,
+)
+from ._admmauxconv import (
     CircularConvolveSolver,
     FBlockCircularConvolveSolver,
     G0BlockCircularConvolveSolver,
